backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pypdf2
openai
python-dotenv
pydantic[email]
aiofiles
httpx
